dz7-dz8/delivery/requirements.txt
asgiref==3.4.1
astroid==2.8.4
Django==3.2.9
gunicorn==20.1.0
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
Pillow==8.4.0
platformdirs==2.4.0
psycopg2==2.9.1
pylint==2.11.1
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
wrapt==1.13.3

djangorestframework~=3.12.4